Removed all *.m4 files from Git
It is unneccessary to keep them in Git because they are can be recreated
by ./autogen.sh.

Additionally, ./autogen.sh sometimes puts in /m4 other versions of these
files than are stored in Git. That can get quite annoying because git
checkout -f (or switching branches / commits) gets back these files,
forcing to do a more lengthy build.

The m4/ directory is placed in .gitignore which supports the fact that
these files are not needed.

This fixes #1152.